refactor: move icrc function
