+++ frontend/src/lib/utils/icrc-transactions.utils.ts
@@ -1,4 +1,5 @@
 import { NANO_SECONDS_IN_MILLISECOND } from "$lib/constants/constants";
+import type { IcrcTransactionsStoreData } from "$lib/stores/icrc-transactions.store";
 import { toastsError } from "$lib/stores/toasts.store";
 import type { Account } from "$lib/types/account";
 import type {
@@ -149,4 +150,33 @@ export const mapIcrcTransaction = ({
       err,
     });
   }
+};
+
+// TODO: use `oldestTxId` instead of sorting and getting the oldest element's id.
+// It seems that the `Index` canister has a bug.
+/**
+ * Returns the oldest transaction id of an Icrc account.
+ *
+ * @param params
+ * @param {Account} params.account
+ * @param {Principal} params.rootCanisterId
+ * @param {IcrcTransactionsStoreData} params.store
+ * @returns {bigint}
+ */
+export const getOldestTxIdFromStore = ({
+  store,
+  canisterId,
+  account,
+}: {
+  store: IcrcTransactionsStoreData;
+  canisterId: Principal;
+  account: Account;
+}): bigint | undefined => {
+  const accountData = store[canisterId.toText()]?.[account.identifier];
+  if (accountData === undefined) {
+    return;
+  }
+  return accountData.transactions.sort((a, b) =>
+    Number(a.transaction.timestamp - b.transaction.timestamp)
+  )[0].id;
 };
