+++ frontend/src/tests/lib/utils/sns-transactions.utils.spec.ts
@@ -1,8 +1,5 @@
 import type { IcrcTransactionsStoreData } from "$lib/stores/icrc-transactions.store";
-import {
-  getOldestTxIdFromStore,
-  isSnsTransactionsCompleted,
-} from "$lib/utils/sns-transactions.utils";
+import { isSnsTransactionsCompleted } from "$lib/utils/sns-transactions.utils";
 import { mockPrincipal } from "../../mocks/auth.store.mock";
 import { createIcrcTransactionWithId } from "../../mocks/icrc-transactions.mock";
 import {
@@ -52,61 +49,6 @@ describe("sns-transactions.utils", () => {
           account: mockSnsSubAccount,
         })
       ).toBe(true);
-    });
-  });
-
-  describe("getOldestTransactionId", () => {
-    const recentTx = {
-      id: BigInt(1234),
-      transaction: {
-        ...transactionFromMainToSubaccount.transaction,
-        timestamp: BigInt(3000),
-      },
-    };
-    const secondTx = {
-      id: BigInt(1235),
-      transaction: {
-        ...transactionFromMainToSubaccount.transaction,
-        timestamp: BigInt(2000),
-      },
-    };
-    const oldestTx = {
-      id: BigInt(1236),
-      transaction: {
-        ...transactionFromMainToSubaccount.transaction,
-        timestamp: BigInt(1000),
-      },
-    };
-
-    it("returns the id of the oldest tx", () => {
-      const rootCanisterId = mockSnsMainAccount.principal;
-      const transactions = [secondTx, oldestTx, recentTx];
-      const store: IcrcTransactionsStoreData = {
-        [rootCanisterId.toText()]: {
-          [mockSnsMainAccount.identifier]: {
-            transactions,
-            completed: false,
-            oldestTxId: BigInt(1234),
-          },
-        },
-      };
-      expect(
-        getOldestTxIdFromStore({
-          store,
-          rootCanisterId,
-          account: mockSnsMainAccount,
-        })
-      ).toBe(oldestTx.id);
-    });
-
-    it("returns undefined if no data is found", () => {
-      expect(
-        getOldestTxIdFromStore({
-          store: {},
-          rootCanisterId: mockSnsMainAccount.principal,
-          account: mockSnsMainAccount,
-        })
-      ).toBeUndefined();
     });
   });
 });
